Fixed loop where resolver-common added a bad import
